Add Healthbar Above Ennemies + Player on the GUI
Add Healthbar handled on Vehicules.cs / HandlingHealthbar method

Add an offset variable inherited on childrens in order to set correctly the healthbar on the top of the enemies GameObject.

The Health value and the GameObject destruction is handled on the TakeDamage() method.

Override the HandlingHealthbar() method in the Player.cs that will point to the bar on the interface instead of instantiating a new one